Update PREAMBLE.md
I wanted to try to propose an alternative preamble with more lyrical tones, taking into account Europe's history, its role in the world and its hopes for the future. In times of crisis (and this is one of them), any society must be able to rely on the solidity of the values on which it is founded. To give in to emotions and concede the field to illiberal forces is to give them a huge advantage in the hearts of citizens, and even to let them find liberal values boring and ineffective. All political principles need emotional support to be consolidated over time. One of the examples of (bad) values that can remain solid even in times of crisis is nationalism: it has a particular effect on the poor, the unemployed, frustrated intellectuals and the declining middle class. All of them, socially humiliated and dissatisfied with their position, find a sense of dignity and pride in belonging to the nation. The second thing that comes to mind is the concept of framing, described as the ability to create a framework that serves our purposes: in political competition, defining the terms of an argument means winning it. Frames are the mental frameworks that determine our view of the world: they cannot be seen or heard, but make up what scientists call the cognitive unconscious, i.e. those mental structures that we cannot perceive through conscious introspection, but only through the effects they produce. They are activated by language, so if you wanted to change the frames, you would first have to change the language: a new way of speaking would have to be created in order to create a new way of thinking, a process that is by no means simple because the reframing process involves bringing to light both beliefs and deeper cognitive modes. Precisely for this reason, it is a mistake to think that metaphorical thinking and emotionality cannot influence rationality in any way, and that facts are sufficient: they are indeed very important, but they must be framed in a mental framework in order to be meaningful. In a sense, the pro-Europeans - in order to keep the European hope alive - should give birth to a European patriotism capable, on the one hand, of consolidating the values of a united Europe and, on the other hand, of redefining the language of patriotism in a pro-European framework, so that they themselves define the terms of discussion with which they will challenge nationalist ideologies: it is not enough to oppose nationalism with the language of nationalism, otherwise we would give it a huge advantage. In a way, the pro-Europeans must attack first, because (in politics as in war) those who can choose the terrain of the battle have a better chance of winning. That is what I have tried to do in this preamble.